feat: utils created for topic plugin
